+++ ksyun/utils_ksyun_callback.go
@@ -1,6 +1,10 @@
 package ksyun
 
 import (
+	"context"
+	"fmt"
+	"sync"
+
 	"github.com/aws/aws-sdk-go/aws/awserr"
 	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
 )
@@ -21,6 +25,22 @@ type callErrorFunc func(d *schema.ResourceData, client *KsyunClient, call ApiCal
 type executeCallFunc func(d *schema.ResourceData, client *KsyunClient, call ApiCall) (*map[string]interface{}, error)
 type afterCallFunc func(d *schema.ResourceData, client *KsyunClient, resp *map[string]interface{}, call ApiCall) error
 type beforeCallFunc func(d *schema.ResourceData, client *KsyunClient, call ApiCall) (bool, error)
+
+type ApiProcess struct {
+	DryRun bool
+	Ctx    context.Context
+	MulNum int
+
+	d      *schema.ResourceData
+	client *KsyunClient
+
+	apiProcessQueue []ApiCall
+	wg              *sync.WaitGroup
+	errCh           chan error
+	concurrentCh    chan struct{}
+	earlyStop       chan struct{}
+	errors          []error
+}
 
 func ksyunApiCall(api []ksyunApiCallFunc, d *schema.ResourceData, meta interface{}) (err error) {
 	if api != nil {
@@ -103,4 +123,84 @@ func ksyunApiCallProcess(api []ApiCall, d *schema.ResourceData, client *KsyunCli
 		}
 	}
 	return err
+}
+
+func (c *ApiCall) RightNow(d *schema.ResourceData, client *KsyunClient, isDryRun bool) error {
+	return ksyunApiCallNew([]ApiCall{*c}, d, client, isDryRun)
+}
+
+func (a *ApiProcess) PutCalls(candidate ...ApiCall) {
+	a.apiProcessQueue = append(a.apiProcessQueue, candidate...)
+}
+
+func (a *ApiProcess) SetD(d *schema.ResourceData) {
+	a.d = d
+}
+func (a *ApiProcess) GetD() *schema.ResourceData {
+	return a.d
+}
+
+func (a *ApiProcess) Client() *KsyunClient {
+	return a.client
+}
+
+func (a *ApiProcess) SetClient(client *KsyunClient) {
+	a.client = client
+}
+
+// NewApiProcess returns a ApiProcess, but now it doesn't support concurrency
+func NewApiProcess(ctx context.Context, d *schema.ResourceData, client *KsyunClient, dryRun bool) ApiProcess {
+	mulNum := 1
+
+	p := ApiProcess{
+		apiProcessQueue: []ApiCall{},
+		d:               d,
+		client:          client,
+		DryRun:          dryRun,
+		Ctx:             ctx,
+	}
+	p.MulNum = mulNum
+	p.errCh = make(chan error, p.MulNum)
+	p.concurrentCh = make(chan struct{}, p.MulNum)
+	p.wg = &sync.WaitGroup{}
+	return p
+}
+
+// ConRun will process ApiProcess concurrently
+func (a *ApiProcess) ConRun() []error {
+	// receive errs
+	go func() {
+		for callErr := range a.errCh {
+			a.errors = append(a.errors, callErr)
+		}
+	}()
+
+	// concurrency api call run
+	for _, call := range a.apiProcessQueue {
+		select {
+		case <-a.Ctx.Done():
+			a.errors = append(a.errors, fmt.Errorf("stop api call early"))
+			return a.errors
+		case <-a.earlyStop:
+			return a.errors
+		default:
+		}
+
+		a.wg.Add(1)
+		a.concurrentCh <- struct{}{}
+		go func(call ApiCall) {
+			defer func() {
+				a.wg.Done()
+				<-a.concurrentCh
+			}()
+			callErr := call.RightNow(a.d, a.client, a.DryRun)
+			a.errCh <- callErr
+		}(call)
+	}
+	a.wg.Wait()
+	return a.errors
+}
+
+func (a *ApiProcess) Run() error {
+	return ksyunApiCallNew(a.apiProcessQueue, a.d, a.client, a.DryRun)
 }